requirements.txt
numpy
python-dotenv
requests
